move MemoryAllocatingTestFixture to tests/mapreduce/common
