+++ src/tritonsort/tests/mapreduce/common/KVPairBufferTest.h
@@ -5,7 +5,7 @@
 
 #include "config.h"
 #include "mapreduce/common/KeyValuePair.h"
-#include "tests/common/MemoryAllocatingTestFixture.h"
+#include "tests/mapreduce/common/MemoryAllocatingTestFixture.h"
 
 class MemoryAllocatorInterface;
