+++ src/tritonsort/tests/mapreduce/common/KVPairFormatReaderTest.h
@@ -5,7 +5,7 @@
 #include <set>
 
 #include "mapreduce/common/FilenameToStreamIDMap.h"
-#include "tests/common/MemoryAllocatingTestFixture.h"
+#include "tests/mapreduce/common/MemoryAllocatingTestFixture.h"
 #include "tests/themis_core/MockWorkerTracker.h"
 
 class BaseBuffer;
