+++ src/tritonsort/tests/mapreduce/common/SampleMetadataKVPairBufferTest.h
@@ -1,7 +1,7 @@
 #ifndef MAPRED_SAMPLE_METADATA_KV_PAIR_BUFFER_TEST_H
 #define MAPRED_SAMPLE_METADATA_KV_PAIR_BUFFER_TEST_H
 
-#include "tests/common/MemoryAllocatingTestFixture.h"
+#include "tests/mapreduce/common/MemoryAllocatingTestFixture.h"
 
 class SampleMetadataKVPairBufferTest : public MemoryAllocatingTestFixture {
   CPPUNIT_TEST_SUITE( SampleMetadataKVPairBufferTest );
